Rename guidance to tour
