+++ lib/ReactViews/Tour/TourIndicator.jsx
